requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
alembic==1.11.1
antlr4-python3-runtime==4.9.3
anyconfig==0.10.1
anyio==3.7.1
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
binaryornot==0.4.4
blinker==1.6.2
build==0.10.0
cachetools==5.3.1
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.4
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.1.0
cookiecutter==2.2.3
cycler==0.11.0
databricks-cli==0.17.7
decorator==5.1.1
docker==6.1.3
dynaconf==3.1.12
entrypoints==0.4
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.95.2
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.40.0
fsspec==2023.6.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphql-core==3.2.3
greenlet==2.0.2
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
httptools==0.6.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.1
kedro==0.18.11
kedro-docker==0.3.1
kedro-viz==6.3.1
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.0
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mlflow==2.4.2
more-itertools==9.1.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
opt-einsum==3.3.0
orjson==3.9.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
pip-tools==6.14.0
platformdirs==3.8.1
plotly==5.15.0
pluggy==1.0.0
prompt-toolkit==3.0.39
protobuf==4.23.4
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.11
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytoolconfig==1.2.5
pytz==2023.3
pywin32==306
PyYAML==6.0
querystring-parser==1.2.4
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rope==1.9.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
secure==0.3.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.18
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
strawberry-graphql==0.194.4
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
toposort==1.10
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.22.0
waitress==2.1.2
watchfiles==0.19.0
watchgod==0.8.2
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.0
